split out interface.py
