+++ docs/source/api_reference/framework/lifecycle/interface.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.lifecycle.interface
